shopen; Refactoring - fixed if statements and varaible names
I fixed the variable names to not conflict with the actual file variable
name. I also fixed the if statements that are searching for a
specific desktop file, because previously the comparaisons would
always fail due to matching the whole path. I fixed this by using
the shell substring removal.